Rename: svg 파일 이름 일부 수정함에 따라 page 내부 import문 수정됨
